refactor(dashboard): split autoresearch.ts god file into 3 modules
879 lines → 3 focused modules:
- autoresearch-state.ts (156): AsyncResource for loops, manual signals
  for detail (race-condition guard), actions, refresh
- autoresearch-form.ts (261): 13 form signals, submit handler,
  StartAutoresearchForm dialog
- autoresearch.ts (510): UI components + main Autoresearch component

Dependency: autoresearch → state, form → state. No cycles.
All exports preserved for external consumers.
